moving i18n to config
